Enable to build ruby_scala with sbt
