+++ .gitignore
@@ -1,0 +1,7 @@
+target
+lib_managed
+src_managed
+project/boot
+project/plugins/project
+.idea
+.idea_modules
