+++ src/main/scala/jp/scala_users/org/ruby_scala/RubyParser.scala
@@ -1,0 +1,9 @@
+package jp.scala_users.org.ruby_scala
+
+import java.io.FileReader
+object RubyParser extends Ruby {
+  def main(args: Array[String]) {
+    val reader = new FileReader(args(0))
+    println(parseAll(program, reader))
+  }
+}
